% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{as.CalDates}
\alias{as.CalDates}
\title{Convert to a CalDates object}
\usage{
as.CalDates(x)
}
\arguments{
\item{x}{One or more calibrated dated in the form of an oxcAARCalibratedDatesList}
}
\value{
A CalDates object
}
\description{
Convert oxcAARCalibratedDatesList to an rcarbon CalDates object. The function is taken and adapted from rcarbon (Bevan, A. and Crema, E.R., 2019).
}
\examples{
\dontrun{
library(oxcAAR)
library(rcarbon)
quickSetupOxcal()
dates <- data.frame(CRA=c(3200,2100,1900), Error=c(35,40,50))
rcaldates <- rcarbon::calibrate(dates$CRA, dates$Error, calCurves=rep("intcal13"))
ocaldates <- oxcAAR::oxcalCalibrate(c(3200,2100,1900),c(35,40,50),c("a","b","c"))
## Convert to rcarbon format
caldates.o <- as.CalDates(ocaldates)
## Comparison plot
plot(rcaldates$grids[[2]]$calBP,rcaldates$grids[[2]]$PrDens,
type="l", col="green", xlim=c(2300,1900))
lines(caldates.o$grids[[2]]$calBP,caldates.o$grids[[2]]$PrDens, col="blue")
legend("topright", legend=c("rcarbon","OxCal"), col=c("green","blue"), lwd=2)
}
}
