% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{Boundary}
\alias{Boundary}
\title{Returns the Oxcal code for a Boundary}
\usage{
Boundary(names)
}
\arguments{
\item{names}{a optional vector of names for the resulting Phases dates. If given, for each name a boundary is returned. If not given, one Boundary without name is returned.}
}
\value{
a string containing the respective Oxcal code
}
\description{
Boundary returns the OxCal code for a Boundary.
For details concerning the Oxcal simulation please consult the help page of Oxcal.
}
