% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{downloadOxcal}
\alias{downloadOxcal}
\title{Download OxCal}
\usage{
downloadOxcal(path = ".", version = "latest")
}
\arguments{
\item{path}{The directory in which OxCal should be stored.}

\item{version}{The OxCal version to download. Defaults to \code{"latest"}.
See \code{testedOxcalVersions()} for fixed releases explicitly tested
with oxcAAR. Other fixed versions can be requested but will generate a
warning because their compatibility with oxcAAR has not been verified.}
}
\description{
Downloads and extracts an OxCal distribution.
}
