% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{format.oxcAARCalibratedDate}
\alias{format.oxcAARCalibratedDate}
\title{Format an oxcAARCalibratedDate}
\usage{
\method{format}{oxcAARCalibratedDate}(x, ...)
}
\arguments{
\item{x}{An \code{\link{oxcAARCalibratedDate}}.}

\item{...}{Unused (kept for S3 consistency).}
}
\value{
A character string (invisibly).
}
\description{
Format an oxcAARCalibratedDate
}
