% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{get_name}
\alias{get_name}
\alias{get_name.default}
\alias{get_name.oxcAARCalibratedDate}
\alias{get_name.oxcAARCalibratedDatesList}
\title{get names (labcodes)}
\usage{
get_name(x)

\method{get_name}{default}(x)

\method{get_name}{oxcAARCalibratedDate}(x)

\method{get_name}{oxcAARCalibratedDatesList}(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate or oxcAARCalibratedDatesList}
}
\value{
a string or a character vector
}
\description{
queries values from date objects
}
\examples{
\dontrun{
x <- oxcalCalibrate(c(5000, 4500, 3000), c(20, 50, 60))
get_name(x)
y <- oxcalCalibrate(5000, 20)[[1]]
get_name(y)
}

}
\seealso{
Other getter functions:
\code{\link[=get_bp]{get_bp()}},
\code{\link[=get_cal_curve]{get_cal_curve()}},
\code{\link[=get_posterior_probabilities]{get_posterior_probabilities()}},
\code{\link[=get_posterior_sigma_ranges]{get_posterior_sigma_ranges()}},
\code{\link[=get_raw_probabilities]{get_raw_probabilities()}},
\code{\link[=get_sigma_ranges]{get_sigma_ranges()}},
\code{\link[=get_std]{get_std()}}
}
\concept{getter functions}
