% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{get_posterior_probabilities}
\alias{get_posterior_probabilities}
\alias{get_posterior_probabilities.default}
\alias{get_posterior_probabilities.oxcAARCalibratedDate}
\alias{get_posterior_probabilities.oxcAARCalibratedDatesList}
\title{get posterior raw probabilities}
\usage{
get_posterior_probabilities(x)

\method{get_posterior_probabilities}{default}(x)

\method{get_posterior_probabilities}{oxcAARCalibratedDate}(x)

\method{get_posterior_probabilities}{oxcAARCalibratedDatesList}(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate or oxcAARCalibratedDatesList}
}
\value{
a list of three data.frames or a list of those lists
}
\description{
queries values from date objects
}
\seealso{
Other getter functions:
\code{\link[=get_bp]{get_bp()}},
\code{\link[=get_cal_curve]{get_cal_curve()}},
\code{\link[=get_name]{get_name()}},
\code{\link[=get_posterior_sigma_ranges]{get_posterior_sigma_ranges()}},
\code{\link[=get_raw_probabilities]{get_raw_probabilities()}},
\code{\link[=get_sigma_ranges]{get_sigma_ranges()}},
\code{\link[=get_std]{get_std()}}
}
\concept{getter functions}
