% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{get_std}
\alias{get_std}
\alias{get_std.default}
\alias{get_std.oxcAARCalibratedDate}
\alias{get_std.oxcAARCalibratedDatesList}
\title{get std values (standard deviations)}
\usage{
get_std(x)

\method{get_std}{default}(x)

\method{get_std}{oxcAARCalibratedDate}(x)

\method{get_std}{oxcAARCalibratedDatesList}(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate or oxcAARCalibratedDatesList}
}
\value{
an integer or a numeric vector
}
\description{
queries values from date objects
}
\examples{
\dontrun{
x <- oxcalCalibrate(c(5000, 4500, 3000), c(20, 50, 60))
get_std(x)
y <- oxcalCalibrate(5000, 20)[[1]]
get_std(y)
}

}
\seealso{
Other getter functions:
\code{\link[=get_bp]{get_bp()}},
\code{\link[=get_cal_curve]{get_cal_curve()}},
\code{\link[=get_name]{get_name()}},
\code{\link[=get_posterior_probabilities]{get_posterior_probabilities()}},
\code{\link[=get_posterior_sigma_ranges]{get_posterior_sigma_ranges()}},
\code{\link[=get_raw_probabilities]{get_raw_probabilities()}},
\code{\link[=get_sigma_ranges]{get_sigma_ranges()}}
}
\concept{getter functions}
