% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDatesList.R
\name{is.oxcAARCalibratedDatesList}
\alias{is.oxcAARCalibratedDatesList}
\title{Checks if a variable is of class oxcAARCalibratedDatesList}
\usage{
is.oxcAARCalibratedDatesList(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"oxcAARCalibratedDatesList"},
  \code{FALSE} otherwise.
}
\description{
Checks if a variable is of class oxcAARCalibratedDatesList
}
