% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDatesList.R
\name{oxcAARCalibratedDatesList}
\alias{oxcAARCalibratedDatesList}
\title{oxcAAR Calibrated Dates List}
\value{
an object of the class 'oxcAARCalibratedDatesList'
}
\description{
A List of \code{\link{oxcAARCalibratedDate}}
}
