% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{oxcal_Sum}
\alias{oxcal_Sum}
\title{Wraps an Oxcal string into a Oxcal sum function}
\usage{
oxcal_Sum(oxcal_string, name = " Sum ")
}
\arguments{
\item{oxcal_string}{The Oxcal script that should be wrapped (vector or single string)}

\item{name}{The name attribute for the resulting sum function}
}
\value{
A new oxcal script as string
}
\description{
Wraps an Oxcal string into a Oxcal sum function
}
