% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcalCalibrate.R
\name{oxcalCalibrate}
\alias{oxcalCalibrate}
\title{Calibrates a 14C date using oxcal}
\usage{
oxcalCalibrate(bp, std, names = 1:length(bp))
}
\arguments{
\item{bp}{A vector containing the bp dates of the measurements}

\item{std}{A vector containing the standard deviations of the measurements}

\item{names}{The names of the measurements, usually the Laboratory numbers}
}
\value{
An object of class \code{\link{oxcAARCalibratedDatesList}}
}
\description{
Calibrates a 14C date using oxcal
}
