% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcalSimulate.R
\name{oxcalSimulate}
\alias{oxcalSimulate}
\title{Simulates 14C dates using oxcal}
\usage{
oxcalSimulate(c_date, std, names = 1:length(c_date))
}
\arguments{
\item{c_date}{A vector containing the calender dates to be simulated}

\item{std}{A vector containing the standard deviations for the simulated dates}

\item{names}{The names of the measurements, usually the Laboratory numbers}
}
\value{
An object of class \code{\link{oxcAARCalibratedDatesList}}
}
\description{
Simulates 14C dates using oxcal
}
