% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{parseFullOxcalOutput}
\alias{parseFullOxcalOutput}
\title{Parses an Oxcal Output File completely into R}
\usage{
parseFullOxcalOutput(output)
}
\arguments{
\item{output}{The output of Oxcal as vector of strings (one string per line).}
}
\value{
A list containing all informations provided by Oxcal as list.
}
\description{
Takes the output of Oxcal as vector of strings (one string per line) and parse it as list.
}

