% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{parseOxcalOutput}
\alias{parseOxcalOutput}
\title{Parses an Oxcal Output File into R}
\usage{
parseOxcalOutput(result, first = FALSE, only.R_Date = TRUE)
}
\arguments{
\item{result}{The output of Oxcal as vector of strings (one string per line).}

\item{first}{Return the first date only}

\item{only.R_Date}{Return the informations for R_Dates}
}
\value{
A list containing all informations provided by Oxcal as list.
}
\description{
Takes the output of Oxcal as vector of strings (one string per line) and parse it as list.
}
