% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{Phase}
\alias{Phase}
\title{Returns the Oxcal code for Phase}
\usage{
Phase(r_dates_strings, names = "")
}
\arguments{
\item{r_dates_strings}{a vector containing strings of OxCal code, usually consisting of R_Date commands, but any other code strings might be used that can be interpreted by OxCal within a Phase}

\item{names}{a optional vector of names for the resulting Phases}
}
\value{
a string containing the respective Oxcal code
}
\description{
Phase takes a set of R_Dates as vectors, and returns
a bit of oxcal code that can be used to feed it into oxcal.
In this code the R_Dates are encapsulated in an OxCal Phases, one Phase for each string.
For details concerning the Oxcal simulation please consult the help page of Oxcal.
}
