% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{plot.oxcAARCalibratedDate}
\alias{plot.oxcAARCalibratedDate}
\title{Plot an oxcAARCalibratedDate}
\usage{
\method{plot}{oxcAARCalibratedDate}(x, use_ggplot = TRUE, ...)
}
\arguments{
\item{x}{An \code{\link{oxcAARCalibratedDate}}.}

\item{use_ggplot}{Logical. If \code{TRUE} (default) and ggplot2 is available,
a ggplot2 plot is produced.}

\item{...}{Further arguments passed to the underlying plotting function.}
}
\description{
Plot an oxcAARCalibratedDate
}
