% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{print.oxcAARCalibratedDate}
\alias{print.oxcAARCalibratedDate}
\title{Print an oxcAARCalibratedDate}
\usage{
\method{print}{oxcAARCalibratedDate}(x, ...)
}
\arguments{
\item{x}{An \code{\link{oxcAARCalibratedDate}}.}

\item{...}{Passed to \code{format()}.}
}
\description{
Print an oxcAARCalibratedDate
}
