% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDatesList.R
\name{print.oxcAARCalibratedDatesList}
\alias{print.oxcAARCalibratedDatesList}
\title{Print an oxcAARCalibratedDatesList}
\usage{
\method{print}{oxcAARCalibratedDatesList}(x, ...)
}
\arguments{
\item{x}{An \code{\link{oxcAARCalibratedDatesList}}.}

\item{...}{Further arguments passed to \code{print()} of single dates.}
}
\description{
Print an oxcAARCalibratedDatesList
}
