% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{quickSetupOxcal}
\alias{quickSetupOxcal}
\title{Quick OxCal setup}
\usage{
quickSetupOxcal(
  os = Sys.info()["sysname"],
  path = tempdir(),
  version = "latest",
  force = FALSE
)
}
\arguments{
\item{os}{The operating system of the workstation. Default: automatic
  determination. Options:
\itemize{
  \item{\bold{Linux}}
  \item{Windows}
  \item{Darwin}
}}

\item{path}{The path to the directory where OxCal is or should be stored.
Default: \code{tempdir()}. For persistent use, installation in a permanent
location is recommended.}

\item{version}{The OxCal version to install. Defaults to \code{"latest"}.
Specific versions can be selected for reproducible installations.
See \code{testedOxcalVersions()} for fixed releases explicitly tested
with oxcAAR.}

\item{force}{Force download and setup even if an existing OxCal
installation is detected.}
}
\description{
Downloads OxCal and sets the executable path correctly.
}
\examples{
\dontrun{
  quickSetupOxcal()
}

}
\author{
Clemens Schmid
}
