% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{R_Date}
\alias{R_Date}
\title{Returns the Oxcal code for the calibration of 14C dates}
\usage{
R_Date(names, r_dates, stds)
}
\arguments{
\item{names}{a vector of names for the dates}

\item{r_dates}{a vector containing the BP dates that should be calibrated}

\item{stds}{a vector containing the standard deviation that should be calibrated (length 1 or same length as r_dates)}
}
\value{
a string containing the respective Oxcal code
}
\description{
R_Date takes names, BP dates and standard deviation for those dates as vectors, and returns
a bit of oxcal code that can be used to feed it into oxcal.
For details concerning the Oxcal calibration please consult the help page of Oxcal.
}
