% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_system_functions.R
\name{readOxcalOutput}
\alias{readOxcalOutput}
\title{Reads the content of the Oxcal js output file}
\usage{
readOxcalOutput(output_file)
}
\arguments{
\item{output_file}{The path to a Oxcal js output file.}
}
\value{
The content of the Oxcal js output file as vector of strings for each line.
}
\description{
Reads the content of the Oxcal js output file as vector of strings for each line.
}
\author{
Martin Hinz
}

