% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxcalCalibratedDate.R
\name{roxcalCalibratedDate}
\alias{roxcalCalibratedDate}
\title{Roxcal Calibratred Dates Object}
\usage{
roxcalCalibratedDate(name, bp, std, cal_curve, sigma_ranges, raw_probabilities)
}
\arguments{
\item{name}{a string giving the name of the date (usually the lab number)}

\item{bp}{a integer giving the BP value for the date}

\item{std}{a integer giving the standard deviation for the date}

\item{cal_curve}{a string giving the used calibration curve}

\item{sigma_ranges}{a list of three elements (one, two, three sigma),
each a data frame with start, end and probability giving}

\item{raw_probabilities}{a data frame of dates and the related probabilities for each date}
}
\value{
an object of the class \code{'roxcalCalibratedDate'}
}
\description{
The function \code{roxcalCalibratedDate} is used to create an object for a calibrated date.
}
