% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxcalCalibratedDatesList.R
\name{roxcalCalibratedDatesList}
\alias{roxcalCalibratedDatesList}
\title{Roxcal Calibrated Dates List}
\value{
an object of the class 'roxcalCalibratedDatesList'
}
\description{
A List of \code{\link{roxcalCalibratedDate}}
}
