% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{Sequence}
\alias{Sequence}
\title{Returns the Oxcal code for Sequence}
\usage{
Sequence(sequence_elements, names = "")
}
\arguments{
\item{sequence_elements}{a vector containing strings of OxCal code, usually consisting of Phase or R_Date commands, but any other code strings might be used that can be interpreted by OxCal within a Sequence}

\item{names}{a optional vector of names for the resulting Sequences}
}
\value{
a string containing the respective Oxcal code
}
\description{
Sequence takes a set of Phases or R_Dates as vectors, and returns
a bit of oxcal code that can be used to feed it into OxCal.
In this code the Phases and/or R_Dates are encapsulated in an OxCal Sequence.
For details concerning the Oxcal simulation please consult the help page of Oxcal.
}
