% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{setOxcalExecutablePath}
\alias{setOxcalExecutablePath}
\title{Set the OxCal executable path}
\usage{
setOxcalExecutablePath(path)
}
\arguments{
\item{path}{The path to the OxCal executable.}
}
\description{
Stores the path to the OxCal executable for use by other oxcAAR functions.
}
\examples{
\dontrun{
setOxcalExecutablePath("/home/martin/Documents/scripte/OxCal/bin/OxCalLinux")
}

}
\author{
Martin Hinz
}
