% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_plot_functions.R
\name{calcurve_plot}
\alias{calcurve_plot}
\title{Plots calibrated dates on the calibration curve}
\usage{
calcurve_plot(calibrated_dates, dates_sigma_ranges = c("two_sigma",
  "one_sigma", "three_sigma"))
}
\arguments{
\item{calibrated_dates}{Either one or a list of calibrated dates}

\item{dates_sigma_ranges}{the sigma ranges used for the error bars}
}
\description{
Plots calibrated dates on the calibration curve
}
