% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_plot_functions.R
\name{calcurve_plot}
\alias{calcurve_plot}
\title{Plots calibrated dates on the calibration curve}
\usage{
calcurve_plot(
  x,
  dates_sigma_ranges = NULL,
  uncal_range = TRUE,
  cal_range = TRUE
)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate or oxcAARCalibratedDatesList}

\item{dates_sigma_ranges}{character. The sigma range used for the error bars ("two_sigma", "one_sigma" or "three_sigma")}

\item{uncal_range}{logical. If TRUE (default), the plot contains error bars for the the uncalibrated age}

\item{cal_range}{logical. If TRUE (default), the plot contains error bars for the the calibrated age}
}
\description{
Plots calibrated dates on the calibration curve
}
