% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeOxcalScript.R
\name{executeOxcalScript}
\alias{executeOxcalScript}
\title{Executes an Oxcal Script}
\usage{
executeOxcalScript(oxcal_script)
}
\arguments{
\item{oxcal_script}{A string containing the Oxcal commands that should be processed.}
}
\value{
The path to the js output file
}
\description{
Takes an Oxcal Script, hands it over to oxcal and receives the output that is read from the output file
}
\author{
Martin Hinz
}
