% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{get_bp}
\alias{get_bp}
\title{get date bp value (age)}
\usage{
get_bp(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate}
}
\value{
an integer
}
\description{
get date bp value (age)
}
\examples{
\dontrun{
x <- oxcalCalibrate(5000, 20)[[1]]
get_bp(x)
}

}
