% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{get_bp}
\alias{get_bp}
\alias{get_bp.default}
\alias{get_bp.oxcAARCalibratedDate}
\alias{get_bp.oxcAARCalibratedDatesList}
\title{get bp values (ages)}
\usage{
get_bp(x)

\method{get_bp}{default}(x)

\method{get_bp}{oxcAARCalibratedDate}(x)

\method{get_bp}{oxcAARCalibratedDatesList}(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate or oxcAARCalibratedDatesList}
}
\value{
an integer or a numeric vector
}
\description{
queries values from date objects
}
\examples{
\dontrun{
x <- oxcalCalibrate(c(5000, 4500, 3000), c(20, 50, 60))
get_bp(x)
y <- oxcalCalibrate(5000, 20)[[1]]
get_bp(y)
}

}
\seealso{
Other getter functions: 
\code{\link{get_cal_curve}()},
\code{\link{get_name}()},
\code{\link{get_raw_probabilities}()},
\code{\link{get_sigma_ranges}()},
\code{\link{get_std}()}
}
\concept{getter functions}
