% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{get_cal_curve}
\alias{get_cal_curve}
\title{get calibration curve name}
\usage{
get_cal_curve(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate}
}
\value{
a string
}
\description{
get calibration curve name
}
\examples{
\dontrun{
x <- oxcalCalibrate(5000, 20)[[1]]
get_cal_curve(x)
}

}
