% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{get_cal_curve}
\alias{get_cal_curve}
\alias{get_cal_curve.default}
\alias{get_cal_curve.oxcAARCalibratedDate}
\alias{get_cal_curve.oxcAARCalibratedDatesList}
\title{get calibration curve names}
\usage{
get_cal_curve(x)

\method{get_cal_curve}{default}(x)

\method{get_cal_curve}{oxcAARCalibratedDate}(x)

\method{get_cal_curve}{oxcAARCalibratedDatesList}(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate or oxcAARCalibratedDatesList}
}
\value{
a string or a character vector
}
\description{
queries values from date objects
}
\examples{
\dontrun{
x <- oxcalCalibrate(c(5000, 4500, 3000), c(20, 50, 60))
get_cal_curve(x)
y <- oxcalCalibrate(5000, 20)[[1]]
get_cal_curve(y)
}

}
\seealso{
Other getter functions: \code{\link{get_bp}},
  \code{\link{get_name}},
  \code{\link{get_raw_probabilities}},
  \code{\link{get_sigma_ranges}}, \code{\link{get_std}}
}
