% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{get_name}
\alias{get_name}
\title{get date name}
\usage{
get_name(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate}
}
\value{
a string
}
\description{
get date name
}
\examples{
\dontrun{
x <- oxcalCalibrate(5000, 20)[[1]]
get_name(x)
}

}
