% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{get_raw_probabilities}
\alias{get_raw_probabilities}
\title{get raw probabilities}
\usage{
get_raw_probabilities(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate}
}
\value{
a data.frame
}
\description{
get raw probabilities
}
\examples{
\dontrun{
x <- oxcalCalibrate(5000, 20)[[1]]
get_raw_probabilities(x)
}

}
