% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{get_raw_probabilities}
\alias{get_raw_probabilities}
\alias{get_raw_probabilities.default}
\alias{get_raw_probabilities.oxcAARCalibratedDate}
\alias{get_raw_probabilities.oxcAARCalibratedDatesList}
\title{get raw probabilities}
\usage{
get_raw_probabilities(x)

\method{get_raw_probabilities}{default}(x)

\method{get_raw_probabilities}{oxcAARCalibratedDate}(x)

\method{get_raw_probabilities}{oxcAARCalibratedDatesList}(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate or oxcAARCalibratedDatesList}
}
\value{
a data.frame or a list of data.frames
}
\description{
queries values from date objects
}
\examples{
\dontrun{
x <- oxcalCalibrate(c(5000, 4500, 3000), c(20, 50, 60))
get_raw_probabilities(x)
y <- oxcalCalibrate(5000, 20)[[1]]
get_raw_probabilities(y)
}

}
\seealso{
Other getter functions: 
\code{\link{get_bp}()},
\code{\link{get_cal_curve}()},
\code{\link{get_name}()},
\code{\link{get_sigma_ranges}()},
\code{\link{get_std}()}
}
\concept{getter functions}
