% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{get_sigma_ranges}
\alias{get_sigma_ranges}
\title{get sigma ranges}
\usage{
get_sigma_ranges(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate}
}
\value{
a list of three data.frames
}
\description{
get sigma ranges
}
\examples{
\dontrun{
x <- oxcalCalibrate(5000, 20)[[1]]
get_sigma_ranges(x)
}

}
