% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{get_std}
\alias{get_std}
\title{get date std value (standard deviation)}
\usage{
get_std(x)
}
\arguments{
\item{x}{an object of class oxcAARCalibratedDate}
}
\value{
an integer
}
\description{
get date std value (standard deviation)
}
\examples{
\dontrun{
x <- oxcalCalibrate(5000, 20)[[1]]
get_std(x)
}

}
