% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{is.oxcAARCalibratedDate}
\alias{is.oxcAARCalibratedDate}
\title{Checks if a variable is of class oxcAARCalibratedDate}
\usage{
is.oxcAARCalibratedDate(x)
}
\arguments{
\item{x}{a variable}
}
\value{
true if x is a oxcAARCalibratedDate, false otherwise
}
\description{
Checks if a variable is of class oxcAARCalibratedDate
}
