% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDatesList.R
\name{is.oxcAARCalibratedDatesList}
\alias{is.oxcAARCalibratedDatesList}
\title{Checks if a variable is of class oxcAARCalibratedDatesList}
\usage{
is.oxcAARCalibratedDatesList(x)
}
\arguments{
\item{x}{a variable}
}
\value{
true if x is a oxcAARCalibratedDatesList, false otherwise
}
\description{
Checks if a variable is of class oxcAARCalibratedDatesList
}
