% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDate.R
\name{oxcAARCalibratedDate}
\alias{oxcAARCalibratedDate}
\title{oxcAAR Calibrated Dates Object}
\usage{
oxcAARCalibratedDate(name, bp, std, cal_curve, sigma_ranges, raw_probabilities,
  posterior_probabilities = NA, posterior_sigma_ranges = NA)
}
\arguments{
\item{name}{a string giving the name of the date (usually the lab number)}

\item{bp}{a integer giving the BP value for the date}

\item{std}{a integer giving the standard deviation for the date}

\item{cal_curve}{a list containing information about the calibration curve (name, resolution, bp, bc, sigma)}

\item{sigma_ranges}{a list of three elements (one, two, three sigma),
each a data frame with start, end and probability giving}

\item{raw_probabilities}{a data frame of dates and the related probabilities for each date}

\item{posterior_probabilities}{a data frame of dates and the related posterior probabilities for each date}

\item{posterior_sigma_ranges}{a list of three elements (one, two, three sigma),
each a data frame with start, end and probability giving for the posterior probabilities}
}
\value{
an object of the class \code{'oxcAARCalibratedDate'}
}
\description{
The function \code{oxcAARCalibratedDate} is used to create an object for a calibrated date.
}
