% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcAARCalibratedDatesList.R
\name{oxcAARCalibratedDatesList}
\alias{oxcAARCalibratedDatesList}
\title{oxcAAR Calibrated Dates List}
\value{
An object of class \code{"oxcAARCalibratedDatesList"} (a list).
}
\description{
A list of calibrated dates, i.e. objects of class \code{\link{oxcAARCalibratedDate}}.
}
