% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{quickSetupOxcal}
\alias{quickSetupOxcal}
\title{Quick OxCal setup}
\usage{
quickSetupOxcal(os = Sys.info()["sysname"], path = tempdir())
}
\arguments{
\item{os}{The operating system of the workstation. Default: automatic determination. Options:
\itemize{
  \item{\bold{Linux}}
  \item{Windows}
  \item{Darwin}
}}

\item{path}{The path to the directory where Oxcal is or should be stored. Default: "tempdir()".
I recommend thought to install it permanently.}
}
\description{
Downloads the latest version of Oxcal and sets the executable path correctly
}
\examples{
\dontrun{
  quickSetupOxcal()
}

}
\author{
Clemens Schmid
}
