% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{setOxcalExecutablePath}
\alias{setOxcalExecutablePath}
\title{Setting the Oxcal program path for further use}
\usage{
setOxcalExecutablePath(path)
}
\arguments{
\item{path}{The path to the Oxcal executable}
}
\description{
Stores the path to the oxcal executable it in internally for other functions.
}
\examples{
\dontrun{
connectOxcal('/home/martin/Documents/scripte/OxCal/bin/OxCalLinux')
}
}
\author{
Martin Hinz
}

